make regions package
